# Update the ugorji json/msgpack library