--- a/extras/generate_latest_map
+++ b/extras/generate_latest_map
@@ -206,19 +206,19 @@ function generate_latest_map() {
     // in the future.  See https://github.com/weaveworks/scope/pull/1709
     // for more information.
     func (m ${latest_map_type}) CodecEncodeSelf(encoder *codec.Encoder) {
-        z, r := codec.GenHelperEncoder(encoder)
+        _, r := codec.GenHelperEncoder(encoder)
         if m == nil {
             r.EncodeNil()
             return
         }
-        r.EncodeMapStart(m.Size())
+        r.WriteMapStart(m.Size())
         for _, val := range m {
-            z.EncSendContainerState(containerMapKey)
+            r.WriteMapElemKey()
             r.EncodeString(cUTF8, val.key)
-            z.EncSendContainerState(containerMapValue)
+            r.WriteMapElemValue()
             val.CodecEncodeSelf(encoder)
         }
-        z.EncSendContainerState(containerMapEnd)
+        r.WriteMapEnd()
     }
 
     // CodecDecodeSelf implements codec.Selfer.
@@ -227,7 +227,7 @@ function generate_latest_map() {
     // undocumented, internal APIs as for CodecEncodeSelf.
     func (m *${latest_map_type}) CodecDecodeSelf(decoder *codec.Decoder) {
         *m = nil
-        z, r := codec.GenHelperDecoder(decoder)
+        _, r := codec.GenHelperDecoder(decoder)
         if r.TryDecodeAsNil() {
             return
         }
@@ -240,19 +240,19 @@ function generate_latest_map() {
             if length < 0 && r.CheckBreak() {
                 break
             }
-            z.DecSendContainerState(containerMapKey)
+            r.ReadMapElemKey()
             var key string
             if !r.TryDecodeAsNil() {
                 key = r.DecodeString()
             }
             i := m.locate(key)
             (*m)[i].key = key
-            z.DecSendContainerState(containerMapValue)
+            r.ReadMapElemValue()
             if !r.TryDecodeAsNil() {
                 (*m)[i].CodecDecodeSelf(decoder)
             }
         }
-        z.DecSendContainerState(containerMapEnd)
+        r.ReadMapEnd()
     }
 
     // MarshalJSON shouldn't be used, use CodecEncodeSelf instead.