--- a/report/latest_map_generated.go
+++ b/report/latest_map_generated.go
@@ -174,19 +174,19 @@ func (m StringLatestMap) DeepEqual(n StringLatestMap) bool {
 // in the future.  See https://github.com/weaveworks/scope/pull/1709
 // for more information.
 func (m StringLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
-	z, r := codec.GenHelperEncoder(encoder)
+	_, r := codec.GenHelperEncoder(encoder)
 	if m == nil {
 		r.EncodeNil()
 		return
 	}
-	r.EncodeMapStart(m.Size())
+	r.WriteMapStart(m.Size())
 	for _, val := range m {
-		z.EncSendContainerState(containerMapKey)
+		r.WriteMapElemKey()
 		r.EncodeString(cUTF8, val.key)
-		z.EncSendContainerState(containerMapValue)
+		r.WriteMapElemValue()
 		val.CodecEncodeSelf(encoder)
 	}
-	z.EncSendContainerState(containerMapEnd)
+	r.WriteMapEnd()
 }
 
 // CodecDecodeSelf implements codec.Selfer.
@@ -195,7 +195,7 @@ func (m StringLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
 // undocumented, internal APIs as for CodecEncodeSelf.
 func (m *StringLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
 	*m = nil
-	z, r := codec.GenHelperDecoder(decoder)
+	_, r := codec.GenHelperDecoder(decoder)
 	if r.TryDecodeAsNil() {
 		return
 	}
@@ -208,19 +208,19 @@ func (m *StringLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
 		if length < 0 && r.CheckBreak() {
 			break
 		}
-		z.DecSendContainerState(containerMapKey)
+		r.ReadMapElemKey()
 		var key string
 		if !r.TryDecodeAsNil() {
 			key = r.DecodeString()
 		}
 		i := m.locate(key)
 		(*m)[i].key = key
-		z.DecSendContainerState(containerMapValue)
+		r.ReadMapElemValue()
 		if !r.TryDecodeAsNil() {
 			(*m)[i].CodecDecodeSelf(decoder)
 		}
 	}
-	z.DecSendContainerState(containerMapEnd)
+	r.ReadMapEnd()
 }
 
 // MarshalJSON shouldn't be used, use CodecEncodeSelf instead.
@@ -395,19 +395,19 @@ func (m NodeControlDataLatestMap) DeepEqual(n NodeControlDataLatestMap) bool {
 // in the future.  See https://github.com/weaveworks/scope/pull/1709
 // for more information.
 func (m NodeControlDataLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
-	z, r := codec.GenHelperEncoder(encoder)
+	_, r := codec.GenHelperEncoder(encoder)
 	if m == nil {
 		r.EncodeNil()
 		return
 	}
-	r.EncodeMapStart(m.Size())
+	r.WriteMapStart(m.Size())
 	for _, val := range m {
-		z.EncSendContainerState(containerMapKey)
+		r.WriteMapElemKey()
 		r.EncodeString(cUTF8, val.key)
-		z.EncSendContainerState(containerMapValue)
+		r.WriteMapElemValue()
 		val.CodecEncodeSelf(encoder)
 	}
-	z.EncSendContainerState(containerMapEnd)
+	r.WriteMapEnd()
 }
 
 // CodecDecodeSelf implements codec.Selfer.
@@ -416,7 +416,7 @@ func (m NodeControlDataLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
 // undocumented, internal APIs as for CodecEncodeSelf.
 func (m *NodeControlDataLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
 	*m = nil
-	z, r := codec.GenHelperDecoder(decoder)
+	_, r := codec.GenHelperDecoder(decoder)
 	if r.TryDecodeAsNil() {
 		return
 	}
@@ -429,19 +429,19 @@ func (m *NodeControlDataLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
 		if length < 0 && r.CheckBreak() {
 			break
 		}
-		z.DecSendContainerState(containerMapKey)
+		r.ReadMapElemKey()
 		var key string
 		if !r.TryDecodeAsNil() {
 			key = r.DecodeString()
 		}
 		i := m.locate(key)
 		(*m)[i].key = key
-		z.DecSendContainerState(containerMapValue)
+		r.ReadMapElemValue()
 		if !r.TryDecodeAsNil() {
 			(*m)[i].CodecDecodeSelf(decoder)
 		}
 	}
-	z.DecSendContainerState(containerMapEnd)
+	r.ReadMapEnd()
 }
 
 // MarshalJSON shouldn't be used, use CodecEncodeSelf instead.