… vendor/github.com/ugorji/go/codec/0doc.go
… vendor/github.com/ugorji/go/codec/README.md
… vendor/github.com/ugorji/go/codec/binc.go
… vendor/github.com/ugorji/go/codec/cbor.go
… vendor/github.com/ugorji/go/codec/cbor_test.go
… vendor/github.com/ugorji/go/codec/codec_test.go
… vendor/github.com/ugorji/go/codec/codecgen/README.md
… vendor/github.com/ugorji/go/codec/codecgen/gen.go
… vendor/github.com/ugorji/go/codec/codecgen_test.go
… vendor/github.com/ugorji/go/codec/decode.go
… vendor/github.com/ugorji/go/codec/encode.go
… vendor/github.com/ugorji/go/codec/fast-path.generated.go
… vendor/github.com/ugorji/go/codec/fast-path.go.tmpl
… vendor/github.com/ugorji/go/codec/fast-path.not.go
… vendor/github.com/ugorji/go/codec/gen-dec-array.go.tmpl
… vendor/github.com/ugorji/go/codec/gen-dec-map.go.tmpl
… vendor/github.com/ugorji/go/codec/gen-helper.generated.go
… vendor/github.com/ugorji/go/codec/gen-helper.go.tmpl
… vendor/github.com/ugorji/go/codec/gen.generated.go
… vendor/github.com/ugorji/go/codec/gen.go
… vendor/github.com/ugorji/go/codec/helper.go
… vendor/github.com/ugorji/go/codec/helper_internal.go
… vendor/github.com/ugorji/go/codec/helper_not_unsafe.go
… vendor/github.com/ugorji/go/codec/helper_test.go
… vendor/github.com/ugorji/go/codec/helper_unsafe.go
… vendor/github.com/ugorji/go/codec/json.go
… vendor/github.com/ugorji/go/codec/mammoth-test.go.tmpl
… vendor/github.com/ugorji/go/codec/mammoth_generated_test.go
… vendor/github.com/ugorji/go/codec/msgpack.go
… vendor/github.com/ugorji/go/codec/noop.go