I wouldn't ordinarily pick a beta version, but it seems the library is about to make a breaking change to the msgpack encoding.

Benchmarks are about the same, after some work:
Master:
```
BenchmarkReportUnmarshal-2   	       1	1770790307 ns/op	458396816 B/op	 3477329 allocs/op
BenchmarkDecode-2            	     100	  14350934 ns/op	  3005481 B/op	   51906 allocs/op
BenchmarkEncode-2            	     100	  16523432 ns/op	  6290948 B/op	   29857 allocs/op
```

This PR:
```
BenchmarkReportUnmarshal-2   	       1	1608796336 ns/op	458262800 B/op	 3477360 allocs/op
BenchmarkDecode-2            	     100	  15738725 ns/op	  3003764 B/op	   51905 allocs/op
BenchmarkEncode-2            	     100	  16321384 ns/op	  6318272 B/op	   29857 allocs/op
```