@rade 2017-12-12
> I wouldn't ordinarily pick a beta version, but it seems the library is about to make a breaking change to the msgpack encoding.

Why update at all?