@bboreham 2017-12-13
> Why update at all?

I started down this rabbit-hole to try to get rid of the memory allocations from `time.Time.MarshalBinary` encoding; this turns out to be inordinately hard and I wondered if the newer codec library would help.

It didn't, but it did trigger a discussion which fixed the `omitempty` for `NodeSet` etc.  Sadly this went in after the breaking change to `time.Time` encoding. I should be able to work something out with that.

So, this PR is a step along the way to somewhere better; not particularly worth merging today but useful to have on file.