@bboreham 2018-05-29
Yeah, close it: the next ugorji update will have to deal with the change in `Time` encoding. 

Bryan

> On 29 May 2018, at 09:22, Alfonso Acosta <notifications@github.com> wrote:
> 
> @bboreham Are you planning to finish up this work anytime soon? Otherwise I would like to close it (I am cleaning up the lingering PRs)
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.